[Fix] Entity 수정_최종
