--- src/main/java/com/nhnacademy/heukbaekbookshop/coupon/domain/BookCoupon.java
@@ -1,31 +1,0 @@
-package com.nhnacademy.heukbaekbookshop.coupon.domain;
-
-import com.nhnacademy.heukbaekbookshop.book.domain.Book;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "books_coupons")
-public class BookCoupon {
-
-    @Id
-    @Column(name = "coupon_id")
-    private Long couponId;
-
-    @OneToOne
-    @MapsId
-    @JoinColumn(name = "coupon_id", insertable = false, updatable = false)
-    private Coupon coupon;
-
-    @ManyToOne
-    @JoinColumn(name = "book_id")
-    private Book book;
-
-}
