--- src/main/java/com/nhnacademy/heukbaekbookshop/coupon/domain/CouponStatus.java
@@ -1,6 +1,0 @@
-package com.nhnacademy.heukbaekbookshop.coupon.domain;
-
-public enum CouponStatus {
-    ABLE,
-    DISABLE
-}
