--- src/main/java/com/nhnacademy/heukbaekbookshop/coupon/domain/CouponType.java
@@ -1,28 +1,0 @@
-package com.nhnacademy.heukbaekbookshop.coupon.domain;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.validator.constraints.Length;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "coupons_types")
-public class CouponType {
-
-    @Id
-    @Column(name = "coupon_type_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotNull
-    @Column(name = "coupon_type_name")
-    private String name;
-
-}
