--- src/main/java/com/nhnacademy/heukbaekbookshop/coupon/repository/MemberCouponRepository.java
@@ -1,15 +1,0 @@
-package com.nhnacademy.heukbaekbookshop.coupon.repository;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import com.nhnacademy.heukbaekbookshop.coupon.domain.MemberCoupon;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface MemberCouponRepository extends JpaRepository<MemberCoupon, Long> {
-    Page<MemberCoupon> findByMemberId(Pageable pageable, Long memberId);
-}
-
