--- src/main/java/com/nhnacademy/heukbaekbookshop/couponpolicy/domain/DisCountType.java
@@ -1,6 +1,0 @@
-package com.nhnacademy.heukbaekbookshop.couponpolicy.domain;
-
-public enum DisCountType {
-    FIXED,
-    PERCENTAGE
-}
