--- src/main/java/com/nhnacademy/heukbaekbookshop/couponpolicy/dto/PolicyRequest.java
@@ -1,13 +1,0 @@
-package com.nhnacademy.heukbaekbookshop.couponpolicy.dto;
-
-import jakarta.persistence.Entity;
-import lombok.Getter;
-
-//@Entity
-@Getter
-
-public class PolicyRequest {
-
-
-
-}
