--- src/main/java/com/nhnacademy/heukbaekbookshop/couponpolicy/dto/PolicyResponse.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.heukbaekbookshop.couponpolicy.dto;
-
-public class PolicyResponse {
-}
