--- src/main/java/com/nhnacademy/heukbaekbookshop/couponpolicy/service/CouponPolicyService.java
@@ -1,14 +1,0 @@
-package com.nhnacademy.heukbaekbookshop.couponpolicy.service;
-
-import com.nhnacademy.heukbaekbookshop.couponpolicy.dto.PolicyRequest;
-import com.nhnacademy.heukbaekbookshop.couponpolicy.dto.PolicyResponse;
-
-import java.util.List;
-
-public interface CouponPolicyService {
-    PolicyResponse createPolicy(PolicyRequest policyRequest);
-    PolicyResponse getPolicyById(long policyId);
-    List<PolicyResponse> getAllPolicies();
-    PolicyResponse updatePolicy(long policyId, PolicyRequest policyRequest);
-    void deletePolicy(long policyId);
-}
