+++ src/main/java/com/nhnacademy/heukbaekbookshop/couponset/coupon/domain/CouponHistory.java
@@ -1,0 +1,41 @@
+package com.nhnacademy.heukbaekbookshop.couponset.coupon.domain;
+
+import com.nhnacademy.heukbaekbookshop.order.domain.OrderBook;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "coupons_histories")
+public class CouponHistory {
+
+    @Id
+    @Column(name = "coupon_history_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @OneToOne
+    @JoinColumn(name = "member_coupon_id")
+    private MemberCoupon memberCoupon;
+
+    @NotNull
+    @Column(name = "coupon_used_at")
+    private LocalDateTime usedAt;
+
+    @ManyToOne
+    @JoinColumns({
+            @JoinColumn(name = "order_id", referencedColumnName = "order_id"),
+            @JoinColumn(name = "book_id", referencedColumnName = "book_id")
+    })
+    private OrderBook orderBook;
+
+}
