+++ src/main/java/com/nhnacademy/heukbaekbookshop/couponset/coupon/domain/CouponStatus.java
@@ -1,0 +1,6 @@
+package com.nhnacademy.heukbaekbookshop.couponset.coupon.domain;
+
+public enum CouponStatus {
+    ABLE,
+    DISABLE
+}
