+++ src/main/java/com/nhnacademy/heukbaekbookshop/couponset/coupon/domain/MemberCoupon.java
@@ -1,0 +1,46 @@
+package com.nhnacademy.heukbaekbookshop.couponset.coupon.domain;
+
+import com.nhnacademy.heukbaekbookshop.memberset.member.domain.Member;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "members_coupons")
+public class MemberCoupon {
+
+    @Id
+    @Column(name = "member_coupon_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "customer_id")
+    private Member member;
+
+    @ManyToOne
+    @JoinColumn(name = "coupon_id")
+    private Coupon coupon;
+
+    @NotNull
+    @Column(name = "is_coupon_used")
+    private boolean isCouponUsed;
+
+    @NotNull
+    @Column(name = "coupon_created_at")
+    private LocalDateTime createdAt;
+
+    @NotNull
+    @Column(name = "coupon_expiration_at")
+    private LocalDateTime expirationAt;
+
+}
