+++ src/main/java/com/nhnacademy/heukbaekbookshop/couponset/couponpolicy/domain/DisCountType.java
@@ -1,0 +1,6 @@
+package com.nhnacademy.heukbaekbookshop.couponset.couponpolicy.domain;
+
+public enum DisCountType {
+    FIXED,
+    PERCENTAGE
+}
