+++ src/main/java/com/nhnacademy/heukbaekbookshop/couponset/couponpolicy/dto/PolicyRequest.java
@@ -1,0 +1,13 @@
+package com.nhnacademy.heukbaekbookshop.couponset.couponpolicy.dto;
+
+import jakarta.persistence.Entity;
+import lombok.Getter;
+
+//@Entity
+@Getter
+
+public class PolicyRequest {
+
+
+
+}
