+++ src/main/java/com/nhnacademy/heukbaekbookshop/couponset/couponpolicy/dto/PolicyResponse.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.heukbaekbookshop.couponset.couponpolicy.dto;
+
+public class PolicyResponse {
+}
