+++ src/main/java/com/nhnacademy/heukbaekbookshop/couponset/couponpolicy/repository/CouponPolicyRepository.java
@@ -1,0 +1,7 @@
+package com.nhnacademy.heukbaekbookshop.couponset.couponpolicy.repository;
+
+import com.nhnacademy.heukbaekbookshop.couponset.couponpolicy.domain.CouponPolicy;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CouponPolicyRepository extends JpaRepository<CouponPolicy, Long> {
+}
